requirements.txt
bs4
requests
selenium
pdfminer.six